Begin work on 003
It's time to create some libraries!  Yay!

I'm starting with counter and primes modules.  (I think that's the
correct term.)  We'll eventually need to add more, but I can't really
find prime-factors until I can find primes.

But again, as this was mostly about understanding Rust, this is a solid
commit.